SDR: split to cpp/hpp
